Split apis
